refactor: Proofread updates. (#12)
The following nouns were written as proper nouns, however they are not proper nouns:

* naval
* webcam
* apprenticeship

Hover-able was changed to the dictionary word hoverable. Changed sentence in line 13 to begin with "However" instead of "BUT". The word but is used primarily as a conjunction and I think in this case your sentence should start with "However". I did not uppercase "However", so you may want to review that change to ensure it matches the tone you would like to be articulated.